+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 
 Forthcoming
 -----------
+* ...
+
+1.3.1 (2020-09-15)
+------------------
 * [keyop] protect command variables with a mutex, `#41 <https://github.com/kobuki-base/kobuki_core/issues/41>`_
 * [core] ros_ prefixes dropped, custom_logging & raw_data_stream demos, `#40 <https://github.com/kobuki-base/kobuki_core/issues/40>`_
 * [infra] recommended firmware versions 1.1.4, 1.2.0, `#37 <https://github.com/kobuki-base/kobuki_core/issues/37>`_